Rework 104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,9 +7,9 @@
 void fibonacci(void)
 {
 	int i;
-	long a = 1;
-	long b = 2;
-	long c;
+	double a = 1;
+	double b = 2;
+	double c;
 	int first = 1;
 
 	for (i = 0; i < 98; i++)
@@ -22,7 +22,7 @@ void fibonacci(void)
 		{
 			printf(", ");
 		}
-		printf("%ld", a);
+		printf("%f", a);
 		c = a + b;
 		a = b;
 		b = c;
